Bluetooth: BAP: Shell: Move broadcast_sink to own shell
To reduce complexity and size of bap.c, move the broadcast_sink
shell to its own shell file. The overall design and implementation
is the same, but there has been some minor changes to fix coding
style or guidelines issues.

The commands has also been renamed since "broadcast_sink" is now
implied by the `bap_broadcast_sink` subcommand.
