+++ subsys/bluetooth/audio/shell/audio.h
@@ -29,6 +29,7 @@
 #include <zephyr/bluetooth/hci_types.h>
 #include <zephyr/bluetooth/iso.h>
 #include <zephyr/kernel.h>
+#include <zephyr/net_buf.h>
 #include <zephyr/shell/shell.h>
 #include <zephyr/sys/atomic_types.h>
 #include <zephyr/sys/byteorder.h>
@@ -168,6 +169,13 @@ bool bap_usb_can_get_full_sdu(struct shell_stream *sh_stream);
 void bap_usb_get_frame(struct shell_stream *sh_stream, enum bt_audio_location chan_alloc,
 		       int16_t buffer[]);
 size_t bap_usb_get_frame_size(const struct shell_stream *sh_stream);
+void bap_broadcast_sink_foreach_stream(void (*func)(struct shell_stream *sh_stream, void *data),
+				       void *data);
+
+void bap_stream_recv_cb(struct bt_bap_stream *stream, const struct bt_iso_recv_info *info,
+			struct net_buf *buf);
+void bap_stream_started_cb(struct bt_bap_stream *bap_stream);
+void bap_stream_stopped_cb(struct bt_bap_stream *stream, uint8_t reason);
 
 struct broadcast_source {
 	bool is_cap: 1;
@@ -188,7 +196,6 @@ struct broadcast_sink {
 	struct bt_le_per_adv_sync *pa_sync;
 	uint8_t received_base[UINT8_MAX];
 	uint8_t base_size;
-	size_t stream_cnt;
 	bool syncable;
 };
